requirements.txt
alpaca-py>=0.28.0
anthropic>=0.34.0
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.26.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.0
python-dateutil>=2.8.2
pytz>=2024.1
streamlit>=1.35.0
plotly>=5.0.0
